requirements.txt
Flask
requests
mysql-connector-python
beautifulsoup4
random-password-generator
python-dotenv
gunicorn
Pillow